+++ CHANGELOG.md
@@ -6,11 +6,14 @@ Version changes are pinned to SDK releases.
 ## [Unreleased]
 
 - general: Rename margin parameters to be more intuitive from `optionBase` to `optionDynamic`. ([#44](https://github.com/zetamarkets/sdk/pull/44))
-- exchange: Fix race condition on `Exchange.load` that would result in NaN margin requirements temporarily. ([#44444ttps://github.com/zetamarkets/sdk/pull/44))
+- exchange: Fix race condition on `Exchange.load` that would result in NaN margin requirements temporarily. ([#44](https://github.com/zetamarkets/sdk/pull/44))
+- events: Trade event now emits `orderId` and `clientOrderId` if set, otherwise it is 0. These are represented in BN. ([#45](https://github.com/zetamarkets/sdk/pull/45))
 - client: Deposit limit for non white-listed users. ([#40](https://github.com/zetamarkets/sdk/pull/40))
 
 ### Breaking
 
+- events: TradeEvent now emits `costOfTrades` instead of `price`. Users can use `utils.getTradeEventPrice(event)` to get the trade price. If your order was a taker and traded against multiple orders in the one insert, the TradeEvent will aggregate across each execution. As a result, the trade price returned is the average trade price across the total taker size. ([#45](https://github.com/zetamarkets/sdk/pull/45))
+- events: TradeEvent size is now a BN. ([#45](https://github.com/zetamarkets/sdk/pull/45))
 - client: Users can only deposit up to a threshold where their balance + unrealized_pnl + new deposit < deposit limit unless they are previously whitelisted. Deposit limit can be found in the `Exchange.state` account. ([#40](https://github.com/zetamarkets/sdk/pull/40))
 
 ## [0.9.2] 2021-12-23